Refactor scan modules: rename scan.py and extract shared code to libscan
Rename each scan module's scan.py to match its module name
(alascan.py, rnascan.py, dnascan.py) and factor the code the modules
duplicated into libscan.py:

- Nucleic-acid mutation helpers shared by rnascan/dnascan: ring/anchor
  atom constants, norm_atom_name, get_atoms_to_keep (parametrized by
  backbone/purines/pyrimidines) and validate_scan_bases (parametrized
  by allowed set).
- filter_interface / build_resname_dict: the identical interface
  user-filtering block used by all three modules' run().
- BaseInterfaceScanner: the shared InterfaceScanner.__init__ setup.
- ModelPointMutation: the identical point-mutation job + scoring flow
  shared by alascan/rnascan; each module keeps a thin run() that
  forwards its module-level mutate/calc_score (kept patchable in tests).

Also apply small cleanups to dnascan (defaultdict, .get, collapsed
branches, RES_CODES literal, cached get_atoms_to_keep per residue,
halved find_base_pairs distance computations) and rnascan (derive
DEFAULT_SCAN_BASES, correct atoms_to_keep annotation, stream file).

dnascan's ModelBasePairMutation (double/two-pass mutation) is genuinely
different and stays in the module. No behavior change; 83 unit tests
across the three modules pass.
