+++ src/haddock/libs/libscan.py
@@ -9,18 +9,229 @@ mutated, how the interface is scanned, and how mutation jobs are executed).
 
 import io
 import os
+import shutil
 from contextlib import redirect_stdout
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 import numpy as np
 import pandas as pd
 
 from haddock import log
 from haddock.clis import cli_score
+from haddock.core.exceptions import ConfigurationError
 from haddock.libs.libontology import PDBFile
 from haddock.libs.libplots import make_alascan_plot, make_rnascan_plot
+
+
+# ---------------------------------------------------------------------------
+# Nucleic-acid mutation helpers (shared by the rnascan and dnascan modules)
+# ---------------------------------------------------------------------------
+# The two nucleic-acid scan modules mutate bases the same way; only the residue
+# naming (one-letter RNA vs two-letter DNA) and the backbone atom set (RNA keeps
+# the 2'-hydroxyl O2') differ. The chemistry-invariant pieces live here so both
+# modules share a single implementation.
+
+# Base ring atoms shared by both purines. Preserving them when mutating one
+# purine into the other keeps the base plane and glycosidic orientation, so CNS
+# only has to rebuild the differing substituents.
+PURINE_BASE_ATOMS = ["N9", "C8", "N7", "C5", "C4", "N3", "C2", "N1", "C6"]
+
+# Base ring atoms shared by both pyrimidines. Same rationale as for the purines.
+PYRIMIDINE_BASE_ATOMS = ["N1", "C2", "O2", "N3", "C4", "C5", "C6"]
+
+# Glycosidic-region anchor atoms kept (and renamed) on cross-type mutations
+# (purine <-> pyrimidine). Because the purine and pyrimidine ring systems do not
+# share atom names, these three atoms are preserved and renamed to their
+# counterpart in the target ring so that CNS rebuilds the new base with the
+# correct glycosidic orientation. The correspondence is:
+#   pyrimidine N1 <-> purine N9   (glycosidic nitrogen)
+#   pyrimidine C2 <-> purine C4
+#   pyrimidine C6 <-> purine C8
+_PYRIMIDINE_TO_PURINE_ANCHORS = {"N1": "N9", "C2": "C4", "C6": "C8"}
+_PURINE_TO_PYRIMIDINE_ANCHORS = {"N9": "N1", "C4": "C2", "C8": "C6"}
+
+
+def norm_atom_name(atom_name: str) -> str:
+    """Normalise atom names so that primes are always written as `'`."""
+    return atom_name.replace("*", "'")
+
+
+def get_atoms_to_keep(
+    ori_resname: str,
+    target_resname: str,
+    *,
+    backbone_atoms: List[str],
+    purines: Tuple[str, ...],
+    pyrimidines: Tuple[str, ...],
+) -> Dict[str, str]:
+    """Return the atoms to preserve when mutating one nucleic-acid base into
+    another.
+
+    The result maps each atom name to keep (as found in the original residue) to
+    the atom name it must be written with in the mutated residue. Backbone atoms
+    and same-ring-type base atoms keep their name (mapped to themselves).
+
+    The sugar-phosphate backbone (``backbone_atoms``) is always kept. When the
+    original and target bases are of the same ring type (both purines or both
+    pyrimidines) the base ring atoms common to that type are kept as well, so
+    that the base orientation is preserved and CNS only rebuilds the differing
+    substituents. For cross-type mutations (purine <-> pyrimidine) the three
+    glycosidic-region anchor atoms are kept and renamed to their counterpart in
+    the target ring system.
+
+    Parameters
+    ----------
+    ori_resname : str
+        Original (wild-type) residue name.
+    target_resname : str
+        Target base residue name.
+    backbone_atoms : list of str
+        Sugar-phosphate backbone atoms to always keep (RNA keeps ``O2'``, DNA
+        does not).
+    purines, pyrimidines : tuple of str
+        Residue names classified as purines / pyrimidines for this module.
+
+    Returns
+    -------
+    dict of str -> str
+        Mapping of original atom name -> atom name to write for the mutated
+        nucleotide.
+    """
+    # Backbone is always kept, with unchanged atom names.
+    atoms_to_keep: Dict[str, str] = {atom: atom for atom in backbone_atoms}
+    if ori_resname in purines and target_resname in purines:
+        atoms_to_keep.update({atom: atom for atom in PURINE_BASE_ATOMS})
+    elif ori_resname in pyrimidines and target_resname in pyrimidines:
+        atoms_to_keep.update({atom: atom for atom in PYRIMIDINE_BASE_ATOMS})
+    elif ori_resname in pyrimidines and target_resname in purines:
+        atoms_to_keep.update(_PYRIMIDINE_TO_PURINE_ANCHORS)
+    elif ori_resname in purines and target_resname in pyrimidines:
+        atoms_to_keep.update(_PURINE_TO_PYRIMIDINE_ANCHORS)
+    return atoms_to_keep
+
+
+def validate_scan_bases(
+    scan_bases: List[str],
+    allowed: Tuple[str, ...],
+    *,
+    base_kind: str,
+    allowed_note: str,
+) -> List[str]:
+    """Validate and normalise a list of target nucleic-acid bases.
+
+    Entries are upper-cased and de-duplicated (keeping first-seen order) and
+    must belong to ``allowed``.
+
+    Parameters
+    ----------
+    scan_bases : list of str
+        Target bases requested by the user.
+    allowed : tuple of str
+        Canonical residue names accepted by the calling module.
+    base_kind : str
+        Human-readable base family (e.g. ``"RNA"`` or ``"DNA"``) used in the
+        "empty list" error message.
+    allowed_note : str
+        Sentence describing the accepted values (and why the other naming is
+        rejected) used in the "invalid entry" error message.
+
+    Returns
+    -------
+    list of str
+        Normalised, de-duplicated list of canonical residue names, in the order
+        first seen.
+
+    Raises
+    ------
+    ConfigurationError
+        If ``scan_bases`` is empty or contains an unrecognised base.
+    """
+    if not scan_bases:
+        raise ConfigurationError(
+            f"'scan_bases' must contain at least one {base_kind} base "
+            f"among {', '.join(allowed)}."
+        )
+    normalised: List[str] = []
+    for base in scan_bases:
+        canonical = str(base).strip().upper()
+        if canonical not in allowed:
+            raise ConfigurationError(
+                f"Invalid 'scan_bases' entry {base!r}. Allowed values are the "
+                f"{allowed_note}"
+            )
+        if canonical not in normalised:
+            normalised.append(canonical)
+    return normalised
+
+
+def filter_interface(
+    interface: Dict[str, List[int]],
+    filter_resdic: Dict[str, List[int]],
+    user_chains: List[str],
+) -> Dict[str, List[int]]:
+    """Restrict the detected interface to the user-selected residues/chains.
+
+    If the user provided target residues (``resdic_*``), the interface is
+    intersected with them. Otherwise, if the user provided target chains, the
+    interface is restricted to those chains. With neither, the interface is
+    returned unchanged.
+
+    Parameters
+    ----------
+    interface : dict
+        Mapping of chain id to its list of interface residue numbers.
+    filter_resdic : dict
+        Mapping of chain id to user-requested residue numbers. The sentinel
+        ``{"_": []}`` means "no residue restriction".
+    user_chains : list of str
+        User-requested chain ids (used only when no residue restriction is set).
+
+    Returns
+    -------
+    dict
+        The filtered interface.
+    """
+    # if user defined target residues, check they are in the interface
+    if filter_resdic != {"_": []}:
+        filtered_interface = {}
+        for chain in filter_resdic:
+            if chain in interface:
+                # intersection of user queried residues and interface residues
+                user_res_valid = list(
+                    set(filter_resdic[chain]).intersection(set(interface[chain]))
+                )
+                # keep the chain only if at least one residue survives
+                if user_res_valid:
+                    filtered_interface[chain] = user_res_valid
+        return filtered_interface
+    # if (user defined target chains) & (no user target residues) - use them
+    if user_chains:
+        return {chain: res for chain, res in interface.items() if chain in user_chains}
+    return interface
+
+
+def build_resname_dict(coords: Dict[Tuple, object]) -> Dict[str, str]:
+    """Map ``"{chain}-{resid}"`` to its residue name from a coordinate dict.
+
+    Parameters
+    ----------
+    coords : dict
+        Coordinate dictionary keyed by ``(chain, resid, atom_name, resname)``
+        tuples, as returned by ``load_coords(..., add_resname=True)``.
+
+    Returns
+    -------
+    dict
+        Mapping of ``"{chain}-{resid}"`` to the residue name (first occurrence).
+    """
+    resname_dict: Dict[str, str] = {}
+    for chain, resid, _atom, resname in coords.keys():
+        key = f"{chain}-{resid}"
+        if key not in resname_dict:
+            resname_dict[key] = resname
+    return resname_dict
 
 
 def make_scan_plot(
@@ -722,3 +933,192 @@ def group_scan_by_cluster(
             entry["delta_bsa"].append(delta_scores[4])
             entry["frac_pr"] += 1
     return clt_scan, clt_pops
+
+
+# ---------------------------------------------------------------------------
+# Shared interface-scan job infrastructure
+# ---------------------------------------------------------------------------
+
+
+class BaseInterfaceScanner:
+    """Common state shared by every module's ``InterfaceScanner``.
+
+    Subclasses add their own mutation target (e.g. ``mutation_res`` for alascan
+    or ``scan_bases`` for the nucleic-acid modules) and implement ``run()``,
+    which detects the interface and builds the module-specific mutation jobs.
+    """
+
+    def __init__(
+        self,
+        model: Union[str, Path, Any],
+        params: Optional[Dict[str, Any]] = None,
+    ) -> None:
+        """Populate the attributes shared by all InterfaceScanner variants.
+
+        Parameters
+        ----------
+        model : str, Path, or model object
+            HADDOCK ``PDBFile`` model object or a path to a PDB file.
+        params : dict, optional
+            Module parameters (interface cutoff, chains, resdic filters,
+            ligand topology/parameter files, ...).
+        """
+        self.model = model
+        self.params = params or {}
+        self.point_mutations_jobs: List[Any] = []
+        self.ligand_param_fname = self.params.get("ligand_param_fname", "")
+        self.ligand_top_fname = self.params.get("ligand_top_fname", "")
+        self.filter_resdic = {
+            key[-1]: value
+            for key, value in self.params.items()
+            if key.startswith("resdic")
+        }
+        if isinstance(model, PDBFile):
+            self.model_path = model.rel_path
+            self.model_id = model.file_name.removesuffix(".pdb")
+        else:
+            # model given as a plain path
+            self.model_path = Path(model)
+            self.model_id = self.model_path.stem
+
+
+class ModelPointMutation:
+    """Execute a single point mutation and score it against the native model.
+
+    Shared by the alascan and rnascan modules, whose per-model scoring flow is
+    identical; only the ``mutate`` primitive differs (protein residue vs RNA
+    base). Each module keeps a thin ``run()`` that forwards its own module-level
+    ``mutate`` and ``calc_score`` to :meth:`_run` (so both stay patchable in the
+    module namespace for tests).
+    """
+
+    def __init__(
+        self,
+        model_path: Path,
+        model_id: str,
+        chain: str,
+        resid: int,
+        ori_resname: str,
+        target_resname: str,
+        native_scores: Tuple[float, float, float, float, float],
+        output_mutants: bool = False,
+        ligand_param_fname: Union[Path, str] = "",
+        ligand_top_fname: Union[Path, str] = "",
+    ) -> None:
+        """Initialize a single point mutation job.
+
+        Parameters
+        ----------
+        model_path : Path
+            Path to the PDB file.
+        model_id : str
+            Identifier for the model.
+        chain : str
+            Chain identifier.
+        resid : int
+            Residue number.
+        ori_resname : str
+            Original residue name.
+        target_resname : str
+            Target residue/base name for the mutation.
+        native_scores : tuple
+            Native model scores (score, vdw, elec, desolv, bsa).
+        output_mutants : bool
+            Whether to keep the mutant PDB files.
+        ligand_param_fname, ligand_top_fname : Union[Path, str]
+            Additional CNS parameter/topology files.
+        """
+        self.model_path = Path(model_path)
+        self.model_id = model_id
+        self.chain = chain
+        self.resid = resid
+        self.ori_resname = ori_resname
+        self.target_resname = target_resname
+        self.native_scores = native_scores
+        self.output_mutants = output_mutants
+        self.ligand_param_fname = ligand_param_fname
+        self.ligand_top_fname = ligand_top_fname
+
+    def _run(self, mutate_func, calc_score_func) -> MutationResult:
+        """Run the mutation with the given ``mutate``/``calc_score`` callables.
+
+        Parameters
+        ----------
+        mutate_func : callable
+            ``mutate(model_path, chain, resid, target_resname) -> Path`` that
+            writes the mutated PDB and returns its path.
+        calc_score_func : callable
+            The module's ``calc_score`` (kept as an argument so it stays
+            patchable in the module namespace during tests).
+        """
+        mutation_id = f"{self.model_id}_{self.chain}{self.resid}{self.target_resname}"
+
+        try:
+            # Setup working directory
+            sc_dir = f"haddock3-score-{mutation_id}"
+            os.makedirs(sc_dir, exist_ok=True)
+
+            # Perform point mutation on pdb file
+            mut_pdb = mutate_func(
+                self.model_path, self.chain, self.resid, self.target_resname
+            )
+
+            # Calculate mutant scores
+            mutant_scores = calc_score_func(
+                mut_pdb,
+                run_dir=sc_dir,
+                outputpdb=self.output_mutants,
+                ligand_param_fname=self.ligand_param_fname,
+                ligand_top_fname=self.ligand_top_fname,
+            )
+
+            # Calculate deltas (native - mutant)
+            n_score, n_vdw, n_elec, n_des, n_bsa = self.native_scores
+            m_score, m_vdw, m_elec, m_des, m_bsa = mutant_scores
+            delta_scores = (
+                n_score - m_score,
+                n_vdw - m_vdw,
+                n_elec - m_elec,
+                n_des - m_des,
+                n_bsa - m_bsa,
+            )
+
+            # Handle output files
+            em_mut_pdb = Path(f"{mut_pdb.stem}_hs.pdb")
+            if not self.output_mutants:
+                # if output_mutants = False, then remove both files
+                if os.path.exists(mut_pdb):
+                    os.remove(mut_pdb)
+                if em_mut_pdb.exists():
+                    os.remove(em_mut_pdb)
+            else:
+                # otherwise keep energy-minimized pdb
+                if os.path.exists(em_mut_pdb):
+                    shutil.move(em_mut_pdb, mut_pdb)
+            # clean up scoring dir
+            if os.path.exists(sc_dir):
+                shutil.rmtree(sc_dir)
+
+            return MutationResult(
+                model_id=self.model_id,
+                chain=self.chain,
+                resid=self.resid,
+                ori_resname=self.ori_resname,
+                target_resname=self.target_resname,
+                mutant_scores=mutant_scores,
+                delta_scores=delta_scores,
+                success=True,
+            )
+
+        except Exception as e:
+            return MutationResult(
+                model_id=self.model_id,
+                chain=self.chain,
+                resid=self.resid,
+                ori_resname=self.ori_resname,
+                target_resname=self.target_resname,
+                mutant_scores=(0, 0, 0, 0, 0),
+                delta_scores=(0, 0, 0, 0, 0),
+                success=False,
+                error_msg=str(e),
+            )
