+++ src/haddock/modules/analysis/rnascan/__init__.py
@@ -69,7 +69,7 @@ from haddock.core.defaults import MODULE_DEFAULT_YAML
 from haddock.libs.libparallel import GenericTask
 from haddock.modules import BaseHaddockModule, get_engine
 from haddock.modules.analysis import get_analysis_exec_mode
-from haddock.modules.analysis.rnascan.scan import (
+from haddock.modules.analysis.rnascan.rnascan import (
     AddDeltaBFactor,
     ClusterOutputer,
     group_scan_by_cluster,
